requirements.txt
python-dotenv
openai==1.68.2
pandas
streamlit > 1.43